srcs/backend/requirements.txt
asgiref==3.8.1
Django==4.2.11
django-extensions==3.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg==3.1.18
psycopg2-binary==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.11.0
